Improve performance (reduce about 200ms for medium size of image)

+++ src/main.rs
@@ -34,7 +34,7 @@ fn main() -> Result<()> {
     })
     .map(|code| {
         if let Some(code) = code {
-            std::process::exit(code as i32)
+            std::process::exit(code)
         }
     })
 }
